+++ extensions/ql-vscode/src/query-history/store/query-history-local-query-domain-mapper.ts
@@ -109,6 +109,11 @@ function mapInitialQueryInfoToDto(
     },
     start: localQueryInitialInfo.start,
     id: localQueryInitialInfo.id,
+    outputDir: localQueryInitialInfo.outputDir
+      ? {
+          querySaveDir: localQueryInitialInfo.outputDir.querySaveDir,
+        }
+      : undefined,
   };
 }